snv filtering cleaned
