+++ configs/base.config
@@ -37,6 +37,24 @@ params {
     cut_het_low = 0.15
 }
 
+// added for snv QC step //
+params {
+    differentialMissingness {
+        mperm = 10000
+        cut = 0.05
+    }
+    hwe {
+        cut = 0.05
+    }
+    missingness {
+        maxRate {
+            cases = 0.01
+            controls = 0.01
+        }
+    }
+    minMAF = 0.05
+}
+
 // added for basic quality control step //
 params {
     cut_diff_miss = "0.05"
